requirements.txt

parse
Pillow